Fix example in docs for translatable scopes
Since `_()` is evaluated at import time the translation would always be the default translation of the Django installation as of process startup time. With `ugettext_lazy` evaluation is delayed until the translated strings are forced to strings in the template, so the current language of the request will be used.